Reverted WCF target to original before multicast modification
